Add integration tests for rollback_dnb_company_updates command
