+++ changelog/company/integration-test-rollback-dnb-updates.feature.md
@@ -1,0 +1,2 @@
+Added integration tests for the `rollback_dnb_company_updates` management command.
+This can now be used as it has been fully tested.
