+++ datahub/dbmaintenance/test/commands/test_rollback_dnb_company_updates.py
@@ -2,25 +2,41 @@ from io import BytesIO
 from unittest import mock
 
 import pytest
+import reversion
 from django.core.management import call_command
+from reversion.models import Version
 
+from datahub.company.models import Company
 from datahub.company.test.factories import CompanyFactory
+from datahub.dnb_api.constants import ALL_DNB_UPDATED_MODEL_FIELDS
+from datahub.dnb_api.utils import update_company_from_dnb
+
 
 pytestmark = pytest.mark.django_db
 
 
-@mock.patch(
-    'datahub.dbmaintenance.management.commands.rollback_dnb_company_updates.'
-    'rollback_dnb_company_update',
-)
-def test_run(mocked_rollback_dnb_company_update, s3_stubber):
+def test_run(s3_stubber, formatted_dnb_company):
     """
-    Test that the command calls the rollback utility for the specified records.
+    Test that the command successfully rolls back the specified records.
     """
-    companies = [
-        CompanyFactory(duns_number='123456789'),
-        CompanyFactory(duns_number='223456789'),
-    ]
+    with reversion.create_revision():
+        companies = [
+            CompanyFactory(duns_number='123456789'),
+            CompanyFactory(duns_number='223456789'),
+        ]
+
+    original_companies = {company.id: Company.objects.get(id=company.id) for company in companies}
+
+    update_descriptor = 'foobar'
+    # Simulate updating the companies from DNB, which sets a revision with the specified
+    # update descriptor
+    for company in companies:
+        formatted_dnb_company['duns_number'] = company.duns_number
+        update_company_from_dnb(
+            company,
+            formatted_dnb_company,
+            update_descriptor=update_descriptor,
+        )
 
     bucket = 'test_bucket'
     object_key = 'test_key'
@@ -41,7 +57,6 @@ def test_run(mocked_rollback_dnb_company_update, s3_stubber):
         },
     )
 
-    update_descriptor = 'foobar'
     call_command(
         'rollback_dnb_company_updates',
         bucket,
@@ -49,8 +64,14 @@ def test_run(mocked_rollback_dnb_company_update, s3_stubber):
         update_descriptor=update_descriptor,
     )
 
+    # Ensure that the companies are reverted to their previous versions
     for company in companies:
-        mocked_rollback_dnb_company_update.assert_any_call(company, update_descriptor)
+        company.refresh_from_db()
+        for field in ALL_DNB_UPDATED_MODEL_FIELDS:
+            assert getattr(company, field) == getattr(original_companies[company.id], field)
+
+        latest_version = Version.objects.get_for_object(company)[0]
+        assert latest_version.revision.comment == f'Reverted D&B update from: {update_descriptor}'
 
 
 @mock.patch(
